requirements.txt
# Core Python
python-dotenv>=1.0.0

# Vector Search
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.2.2
torch>=2.0.0

# LLM / API
requests>=2.31.0

# UI
streamlit>=1.29.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Optional (used during preprocessing / chunking)
langchain>=0.1.0
pydantic>=2.0.0
